+++ Abstandsrechner/src/Rechner.java
@@ -43,8 +43,8 @@ public class Rechner {
 		boolean vielfaches = false;
 
 		for (int i = 0; i < vektor1.Werte.size(); i++) {
-			int wert1 = vektor1.Werte.get(i);
-			int wert2 = vektor2.Werte.get(i);
+			double wert1 = vektor1.Werte.get(i);
+			double wert2 = vektor2.Werte.get(i);
 			if (wert1 == 0 && wert2 == 0) {
 				vielfaches = true;
 			} else if (wert1 != 0 && wert2 != 0) {
@@ -68,10 +68,10 @@ public class Rechner {
 
 		/// GausAlgorithmus;
 
-		List<List<Integer>> zeilen = new ArrayList<>();
+		List<List<Double>> zeilen = new ArrayList<>();
 
 		for (int i = 0; i < gerade.StuetsVectorParameter.Werte.size(); i++) {
-			List<Integer> neueZeile = new ArrayList<Integer>();
+			List<Double> neueZeile = new ArrayList<Double>();
 
 			/// vor dem gleichzeichen
 			neueZeile.add(ebene.RichtungsVectorParameter1.Werte.get(i));
@@ -85,10 +85,10 @@ public class Rechner {
 		}
 
 		// wert mit der die zweite zeile multipliziert werden muss
-		int wert1 = zeilen.get(2).get(0);
+		double wert1 = zeilen.get(2).get(0);
 
 		// wert mit der die dritte zeile mulipliziert werden muss
-		int wert2 = zeilen.get(1).get(0);
+		double wert2 = zeilen.get(1).get(0);
 
 		// multiplizieren zweite und dritte mit dem wert und subtrahiere dann die dritte
 		// zeile von der zweiten.
@@ -138,10 +138,10 @@ public class Rechner {
 
 		// Punkte berechnen:
 		// 3. punkt
-		int punkt3 = zeilen.get(2).get(3) / zeilen.get(2).get(2);
-		int punkt2 = zeilen.get(1).get(3) - zeilen.get(1).get(2) * punkt3;
+		double punkt3 = zeilen.get(2).get(3) / zeilen.get(2).get(2);
+		double punkt2 = zeilen.get(1).get(3) - zeilen.get(1).get(2) * punkt3;
 		punkt2 = punkt2 / zeilen.get(1).get(1);
-		int punkt1 = zeilen.get(0).get(3) - (zeilen.get(0).get(2) * punkt3 + zeilen.get(0).get(1) * punkt2);
+		double punkt1 = zeilen.get(0).get(3) - (zeilen.get(0).get(2) * punkt3 + zeilen.get(0).get(1) * punkt2);
 		punkt1 = punkt1 / zeilen.get(0).get(0);
 
 		for (int i = 0; i < gerade.StuetsVectorParameter.Werte.size(); i++) {
@@ -159,7 +159,7 @@ public class Rechner {
 	private static double BetragBerechnen(Vektor vektor) {
 		double betrag = 0;
 
-		for (int zahl : vektor.Werte) {
+		for (double zahl : vektor.Werte) {
 			betrag = betrag + zahl;
 		}
 		;
@@ -170,7 +170,7 @@ public class Rechner {
 	private static double BetragBerechnenVektor(Vektor vektor) {
 		double betrag = 0;
 
-		for (int zahl : vektor.Werte) {
+		for (double zahl : vektor.Werte) {
 			double quadratzahl = zahl * zahl;
 			betrag = betrag + quadratzahl;
 		}
@@ -217,8 +217,8 @@ public class Rechner {
 		}
 
 		for (int i = 0; i < groeße; i++) {
-			int zwischensumme = hilfsvector1.Werte.get(i + 1) * hilfsvector2.Werte.get(i + 2);
-			int zwischensumme2 = hilfsvector1.Werte.get(i + 2) * hilfsvector2.Werte.get(i + 1);
+			double zwischensumme = hilfsvector1.Werte.get(i + 1) * hilfsvector2.Werte.get(i + 2);
+			double zwischensumme2 = hilfsvector1.Werte.get(i + 2) * hilfsvector2.Werte.get(i + 1);
 
 			normalenVector.Werte.add(zwischensumme - zwischensumme2);
 		}
